Move tooltip selectors to separate files
