+++ scripts/snapshots/es6Files.txt
@@ -154,6 +154,9 @@
   "es6/state/selectors/selectChartOffset.js",
   "es6/state/selectors/selectChartOffsetInternal.js",
   "es6/state/selectors/selectPlotArea.js",
+  "es6/state/selectors/selectTooltipAxis.js",
+  "es6/state/selectors/selectTooltipAxisId.js",
+  "es6/state/selectors/selectTooltipAxisType.js",
   "es6/state/selectors/selectTooltipEventType.js",
   "es6/state/selectors/selectTooltipPayloadSearcher.js",
   "es6/state/selectors/selectTooltipSettings.js",
