+++ scripts/snapshots/libFiles.txt
@@ -154,6 +154,9 @@
   "lib/state/selectors/selectChartOffset.js",
   "lib/state/selectors/selectChartOffsetInternal.js",
   "lib/state/selectors/selectPlotArea.js",
+  "lib/state/selectors/selectTooltipAxis.js",
+  "lib/state/selectors/selectTooltipAxisId.js",
+  "lib/state/selectors/selectTooltipAxisType.js",
   "lib/state/selectors/selectTooltipEventType.js",
   "lib/state/selectors/selectTooltipPayloadSearcher.js",
   "lib/state/selectors/selectTooltipSettings.js",
