+++ scripts/snapshots/typesFiles.txt
@@ -154,6 +154,9 @@
   "types/state/selectors/selectChartOffset.d.ts",
   "types/state/selectors/selectChartOffsetInternal.d.ts",
   "types/state/selectors/selectPlotArea.d.ts",
+  "types/state/selectors/selectTooltipAxis.d.ts",
+  "types/state/selectors/selectTooltipAxisId.d.ts",
+  "types/state/selectors/selectTooltipAxisType.d.ts",
   "types/state/selectors/selectTooltipEventType.d.ts",
   "types/state/selectors/selectTooltipPayloadSearcher.d.ts",
   "types/state/selectors/selectTooltipSettings.d.ts",
