+++ src/state/selectors/selectActivePropsFromChartPointer.ts
@@ -2,11 +2,12 @@ import { createSelector } from 'reselect';
 import { RechartsRootState } from '../store';
 import { ActiveTooltipProps } from '../tooltipSlice';
 import { selectChartLayout } from '../../context/chartLayoutContext';
-import { selectTooltipAxisRangeWithReverse, selectTooltipAxisTicks, selectTooltipAxisType } from './tooltipSelectors';
+import { selectTooltipAxisRangeWithReverse, selectTooltipAxisTicks } from './tooltipSelectors';
 import { selectChartOffsetInternal } from './selectChartOffsetInternal';
 import { combineActiveProps, selectOrderedTooltipTicks } from './selectors';
 import { selectPolarViewBox } from './polarAxisSelectors';
 import { ChartPointer } from '../../util/types';
+import { selectTooltipAxisType } from './selectTooltipAxisType';
 
 const pickChartPointer = (_state: RechartsRootState, chartPointer: ChartPointer) => chartPointer;
