+++ src/state/selectors/selectTooltipAxisId.ts
@@ -1,0 +1,4 @@
+import { RechartsRootState } from '../store';
+import { AxisId } from '../cartesianAxisSlice';
+
+export const selectTooltipAxisId = (state: RechartsRootState): AxisId => state.tooltip.settings.axisId;
