+++ storybook/storybook-addon-recharts/inspectors/TooltipAxisTypeInspector.tsx
@@ -1,5 +1,6 @@
 import { useAppSelector } from '../../../src/state/hooks';
-import { selectTooltipAxisType } from '../../../src/state/selectors/tooltipSelectors';
+
+import { selectTooltipAxisType } from '../../../src/state/selectors/selectTooltipAxisType';
 
 export function TooltipAxisTypeInspector() {
   return useAppSelector(selectTooltipAxisType);
